% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDocumentation.R
\docType{data}
\name{accident_2014}
\alias{accident_2014}
\title{accident_2014}
\format{
A dataset with 30056 observations and 50 variables
}
\source{
/url{http://www.nhtsa.gov/Data/Fatality-Analysis-Reporting-System-(FARS)}
}
\usage{
data(accident_2014)
}
\description{
A dataset reporting accident fatalities data in USA for 2014.
}
\keyword{datasets}
