% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FarsFunctions.R
\name{make_filename}
\alias{make_filename}
\title{Format name and year
This function formats the name of the file with the fars data and the year}
\usage{
make_filename(year)
}
\arguments{
\item{year}{a numeric value indicating the year}
}
\value{
a character string with the formatted file name and year
}
\description{
Format name and year
This function formats the name of the file with the fars data and the year
}
\examples{
make_filename(2013)

}
